pyproject.toml
[tool.poetry]
name = "project2"
version = "0.1.0"
description = "Test Project"
authors = ["Alexander <berpress@gmail.com>"]
packages = [{ include = "gendiff" }]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.2"

[tool.poetry.scripts]
gendiff = "gendiff.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
